requirements.txt
# FastAPI and web framework
fastapi>=0.109.0,<0.115.0
uvicorn[standard]>=0.27.0,<0.31.0
python-multipart>=0.0.9,<0.1.0

# Database
sqlalchemy>=2.0.25,<2.1.0
alembic>=1.13.0,<1.14.0
psycopg2-binary>=2.9.9,<3.0.0

# Async and background tasks
celery>=5.3.6,<5.5.0
redis>=5.0.1,<6.0.0
flower>=2.0.1,<3.0.0

# HTTP client and utilities
httpx>=0.26.0,<0.28.0
aiofiles>=23.2.1,<24.0.0

# Data validation and serialization (Python 3.13 compatible)
pydantic>=2.8.0,<3.0.0
pydantic-settings>=2.4.0,<3.0.0
email-validator>=2.0.0,<3.0.0  # Required for EmailStr validation

# Environment variables
python-dotenv>=1.0.0,<2.0.0

# Authentication and security
passlib>=1.7.4,<2.0.0  # Removed [bcrypt] extra to avoid version conflicts
bcrypt==3.2.0  # Pin to 3.2.0 for passlib compatibility (4.0+ breaks passlib)
python-jose[cryptography]>=3.3.0,<4.0.0

# Development and testing
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<0.24.0
black>=24.0.0,<25.0.0
isort>=5.13.0,<6.0.0

# Logging and monitoring
structlog>=24.1.0,<25.0.0
